+++ bin/utworz-konto-admina.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+python src/manage.py createsuperuser --username admin --noinput --email admin@admin.pl;
+./bin/ustaw-domyslne-haslo-admina.sh