+++ src/zglos_publikacje/urls.py
@@ -8,5 +8,10 @@ urlpatterns = [
         views.Zgloszenie_PublikacjiWizard.as_view(),
         name="nowe_zgloszenie",
     ),
+    path(
+        "edycja_zgloszenia/<uuid:kod_do_edycji>/",
+        views.Zgloszenie_PublikacjiWizard.as_view(),
+        name="edycja_zgloszenia",
+    ),
     path("sukces/", views.Sukces.as_view()),
 ]